Feat : 시그널 구현. 컴파일 시 주석 참고.
