+++ readline.c
@@ -1,0 +1,48 @@
+#include <signal.h>
+#include <stdbool.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <readline/readline.h>
+#include <readline/history.h>
+#include <unistd.h>
+
+void handler(int signum)
+{
+    if (signum != SIGINT)
+        return;
+    write(STDOUT_FILENO, "\n", 1);
+    if (rl_on_new_line() == -1)
+        exit(1);
+    rl_replace_line("", 1);
+    rl_redisplay();
+}
+
+int main(void)
+{
+    int ret;
+    char *line;
+
+    signal(SIGINT, handler);
+    while (true)
+    {
+        line = readline("input> ");
+        if (line)
+        {
+            ret = strcmp(line, "bye");
+            if (ret)
+                printf("output> %s\n", line);
+            add_history(line);
+            free(line);
+            line = NULL;
+            if (!ret)
+                break;
+        }
+        else
+        {
+            printf("d\n");
+            return (1);
+        }
+    }
+    return (0);
+}
