+++ src/install/init_dis.c
@@ -6,7 +6,7 @@
 /*   By: sham <sham@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/03 13:26:04 by sham              #+#    #+#             */
-/*   Updated: 2021/12/03 13:51:25 by sham             ###   ########.fr       */
+/*   Updated: 2021/12/06 11:37:19 by sham             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
