+++ src/install/install_util.c
@@ -6,7 +6,7 @@
 /*   By: sham <sham@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 16:42:35 by sham              #+#    #+#             */
-/*   Updated: 2021/12/01 12:00:10 by sham             ###   ########.fr       */
+/*   Updated: 2021/12/06 11:04:49 by sham             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
