+++ src/signal/get_signal.c
@@ -1,13 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pipe.c                                             :+:      :+:    :+:   */
+/*   get_signal.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sham <sham@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/12/01 13:01:22 by sham              #+#    #+#             */
-/*   Updated: 2021/12/01 17:56:36 by sham             ###   ########.fr       */
+/*   Created: 2021/12/08 11:50:39 by sham              #+#    #+#             */
+/*   Updated: 2021/12/08 12:13:42 by sham             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
